cookpad/requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.5
Markdown==2.6.11
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.0.0
protobuf==3.5.2.post1
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
scipy==1.0.1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
Theano==1.0.1
Werkzeug==0.14.1
